Fix NumPy compatibility and Chang'E 4 lunar day offset bug; bump to v2.1.2
- Replace deprecated np.float with float in chronos.py (removed in NumPy 1.20+)
- Fix ce4lst lunar day off-by-one: start gfposc search 1 day after startdate
  to skip the spurious midnight crossing 6 seconds after Chang'E 4 landing;
  lunarday=1 now works as direct input without any workaround
- Remove lunarday+1 workaround from tests and README; update expected values
  to match current spiceypy output
- Delete stale build/ directory and debug test file (test_20231010.py)
- Add CHANGELOG.md and CLAUDE.md
- Bump version to 2.1.2

+++ time_converter/converters/msl/chronos.py
@@ -42,10 +42,10 @@ def _load_sclk_data(filename=os.path.join(os.path.dirname(__file__), 'msl.tsc'))
         data = np.genfromtxt(filename, dtype=(float, dt.datetime, float, float), skip_header=skip_header, max_rows=max_rows,
                              unpack=False,
                              converters={
-                                 0: np.float,
+                                 0: float,
                                  1: lambda x: dt.datetime.strptime(x.decode('ascii'), "%Y-%jT%H:%M:%S.%f"),
-                                 2: np.float,
-                                 3: np.float
+                                 2: float,
+                                 3: float
                              })
         sclk, utc, dut, sclkrate = data['f0'], data['f1'], data['f2'], data['f3']
         sclk_data = sclk, utc, dut, sclkrate
